+++ src/modwire/layers/app.py
@@ -1,9 +1,0 @@
-from modwire.shared import config
-
-
-class LayersApplication:
-    def __init__(
-        self,
-        config: config.LayersConfig,
-    ):
-        self.config = config