+++ src/modwire/modules/app.py
@@ -1,9 +1,0 @@
-from modwire.shared import config
-
-
-class ModulesApplication:
-    def __init__(
-        self,
-        config: config.ModulesConfig,
-    ):
-        self.config = config